Package s2n_bignum into nix derivation
The scripts in s2n_bignum assume execution within the $ROOT/<arch>/
directory and rely on relative paths. To integrate these scripts with
Nix, it is necessary to adjust the paths used for loading scripts to
ensure compatibility.

Additionally, the Makefile located under $ROOT/<arch> has been copied
to the mlkem-native directory and modified as needed to support this
change.

Signed-off-by: Thing-han, Lim <15379156+potsrevennil@users.noreply.github.com>